Intergration_test/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.2.2
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
dnspython==2.3.0
docx2txt==0.8
email-validator==2.0.0
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==2.2.2
Flask-WTF==1.1.1
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.24.2
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.2
txtai==7.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==2.2.2
WTForms==3.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.11.0
